[#2] HistoryInfoController 테스트를 EmployeeControllerTest로 옮긴다
