+++ src/test/java/kr/rogarithm/econrich/domain/employee/controller/EmployeeControllerTest.java
@@ -12,6 +12,9 @@ import kr.rogarithm.econrich.domain.employee.domain.Employee;
 import kr.rogarithm.econrich.domain.employee.dto.EmployeeResponse;
 import kr.rogarithm.econrich.domain.employee.exception.EmployeeNotFoundException;
 import kr.rogarithm.econrich.domain.employee.service.EmployeeService;
+import kr.rogarithm.econrich.domain.historyInfo.domain.JobHistory;
+import kr.rogarithm.econrich.domain.historyInfo.dto.JobHistoryResponse;
+import kr.rogarithm.econrich.domain.historyInfo.exception.JobHistoryNotFoundException;
 import kr.rogarithm.econrich.domain.historyInfo.service.HistoryInfoService;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -74,4 +77,40 @@ class EmployeeControllerTest {
 
         verify(employeeService).getEmployeeById(id);
     }
+
+    @Test
+    public void getJobHistoryWithInvalidId() throws Exception {
+        Long id = -1L;
+
+        when(historyInfoService.getJobHistoryById(id)).thenThrow(JobHistoryNotFoundException.class);
+
+        this.mockMvc
+                .perform(get("/employees/{employeeId}/historyInfo", id))
+                .andDo(print())
+                .andExpect(status().isNotFound());
+
+        verify(historyInfoService).getJobHistoryById(id);
+    }
+
+    @Test
+    public void getJobHistoryWithValidId() throws Exception {
+        Long id = 101L;
+        JobHistory jobHistory = JobHistory.builder()
+                                          .id(101L)
+                                          .startDate(LocalDate.parse("1989-09-21"))
+                                          .endDate(LocalDate.parse("1993-10-27"))
+                                          .jobId("AC_ACCOUNT")
+                                          .departmentId(110L)
+                                          .build();
+
+        when(historyInfoService.getJobHistoryById(id)).thenReturn(JobHistoryResponse.of(jobHistory));
+
+        this.mockMvc
+                .perform(get("/employees/{employeeId}/historyInfo", id))
+                .andDo(print())
+                .andExpect(status().isOk());
+
+        verify(historyInfoService).getJobHistoryById(id);
+    }
+
 }
